backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic-settings
passlib[bcrypt]
bcrypt<4.0.0
python-jose
python-multipart
aiofiles